Let LightningDataModule return spatiotemporal metadata (#66)
* :card_file_box: Let LightningDataModule return spatiotemporal metadata

Making the LightningDataModule return not only the image, but also spatiotemporal metadata such as the bounding box, coordinate reference system, and date. The bbox and crs is in the raster image's native UTM projection for now, while the date is just a YYYY-MM-DD formatted string. Unit tests have been updated to ensure that the extra metadata is passed through.

* :recycle: Refactor test_geotiffdatapipemodule to use parametrization

Reduce duplicate code py using pytest.mark.parametrize, looping over fit and predict stages.

* :memo: Document returned outputs from _array_to_torch function

Improved the docstring of the _array_to_torch function, mentioning the input parameters (filepath) and the contents of the output dictionary (image, bbox, crs, date). Also updated the type hint of the function.

* :truck: Rename crs to epsg

Since we're storing the EPSG integer and not the CRS representation.