:beers: Implement CLAYModule's predict_step to generate embeddings table
Output embeddings to a geopandas.GeoDataFrame with columns 'source_url', 'date', 'embeddings', and 'geometry'. Essentially copying and adapting the code from a76716409e05f1b9dfb992c373fc6acd1e194bcf in #73, but modifying how the encoder's masking is disabled, and how the mean/average of the embeddings is computed over a slice of the raw embeddings.